Fixes to utils number/string and build watch
