+++ packages/utils/src/index.ts
@@ -7,6 +7,7 @@ export * from './dates';
 export * from './interfaces';
 export * from './logger';
 export * from './misc';
+export * from './numbers';
 export * from './promises';
 export * from './errors';
 export * from './strings';
